requirements.txt
flask==3.0.0
python-dotenv==1.0.0
openai==1.14.0
tiktoken==0.6.0
PyPDF2==3.0.1
python-docx==1.1.0
